refactor(vmscraper): extract fleet stats into stats.go
Move Stats aggregation and version-bucket capping out of scraper.go
into stats.go/stats_test.go, matching the schedule.go layout.

User request: extract stats to a separate file; code partially generated by AI.
